8153086: Conditionally save policy file in Sandbox tests

+++ tests/system/src/test/java/test/util/Util.java
@@ -234,17 +234,24 @@ public class Util {
 
             try {
                 if (workerPatchPolicy != null) {
-                    // with Jake, we need to create a merged java.policy
+                    // with Jigsaw, we need to create a merged java.policy
                     // file that contains the permissions for the Xpatch classes
                     // as well as the permissions needed for this test
-
-                    File tempFile = File.createTempFile("java", "policy");
-                    tempFile.deleteOnExit();
 
                     File wpp = new File(workerPatchPolicy);
                     if (!wpp.exists()) {
                         throw new RuntimeException("Missing workerPatchPolicy");
                     }
+
+                    File tempFile = null;
+                    if (workerDebug) {
+                        tempFile = new File(workerPatchPolicy +
+                                "_" + testAppName);
+                    } else {
+                        tempFile= File.createTempFile("java", "policy");
+                        tempFile.deleteOnExit();
+                    }
+
                     BufferedWriter writer = new BufferedWriter(new FileWriter(tempFile));
 
                     BufferedReader reader1 = new BufferedReader(new FileReader(wpp));
